Carry the idle and busy loop counts as std::int64_t
hpx.max_idle_loop_count and hpx.max_busy_loop_count are read from the
configuration with get_entry_as<std::int64_t>, and they end up being
compared against scheduling_counters' idle_loop_count_ and
busy_loop_count_, which are std::int64_t references. Both ends of that
path already agree.

In between they were passed as std::size_t through
thread_pool_init_parameters, scheduled_thread_pool and the
scheduling_callbacks constructor, which then cast them straight back to
std::int64_t to store them. MSVC reports the first of those conversions
as C4244 on a 32 bit target, and it accounts for 20 of the 1586 warnings
the 32 bit Windows job emits.

Use std::int64_t for the whole path. The two static_casts in
scheduling_callbacks are no longer needed and go away with it.

standalone_thread_pool_executor.cpp already passes
std::numeric_limits<std::int64_t>::max() for both of these, which is
what the values are meant to be and which a 32 bit build was silently
truncating.

Refs #7576

+++ libs/core/thread_pools/include/hpx/thread_pools/detail/scheduling_callbacks.hpp
@@ -27,14 +27,14 @@ namespace hpx::threads::detail {
             background_callback_type&& background = background_callback_type(),
             std::size_t max_background_threads =
                 (std::numeric_limits<std::size_t>::max)(),
-            std::size_t max_idle_loop_count = HPX_IDLE_LOOP_COUNT_MAX,
-            std::size_t max_busy_loop_count = HPX_BUSY_LOOP_COUNT_MAX)
+            std::int64_t max_idle_loop_count = HPX_IDLE_LOOP_COUNT_MAX,
+            std::int64_t max_busy_loop_count = HPX_BUSY_LOOP_COUNT_MAX)
           : outer_(HPX_MOVE(outer))
           , inner_(HPX_MOVE(inner))
           , background_(HPX_MOVE(background))
           , max_background_threads_(max_background_threads)
-          , max_idle_loop_count_(static_cast<std::int64_t>(max_idle_loop_count))
-          , max_busy_loop_count_(static_cast<std::int64_t>(max_busy_loop_count))
+          , max_idle_loop_count_(max_idle_loop_count)
+          , max_busy_loop_count_(max_busy_loop_count)
         {
         }
